minor refactor to move CardWrapper into Card 💥
